Rewrite core g-value handling to make more sense.
Note that GType in glib headers is defined conditionally with:
\#if     GLIB_SIZEOF_SIZE_T != GLIB_SIZEOF_LONG || !defined __cplusplus
typedef gsize                           GType;
\#else   /* for historic reasons, C++ links against gulong GTypes */
typedef gulong                          GType;
\#endif

gsize is defined as unsigned int, so I gave it that type. This, and some other things, should really be
done using cffi-grovel, but somehow I don't think that adding a dependency on a C compiler would go
very well.